chat/pyproject.toml
[tool.poetry]
name = "chat"
version = "0.1.0"
description = ""
authors = ["Kylemerian <lawerenceses@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.112.2"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
jinja2 = "^3.1.4"
itsdangerous = "^2.2.0"
python-multipart = "^0.0.9"
sqlalchemy = "^2.0.34"
asyncpg = "^0.29.0"
databases = "^0.9.0"
psycopg2-binary = "^2.9.9"
fastapi-users = "^13.0.0"
pyjwt = "2.8.0"
typing-extensions = "^4.12.2"
websocket = "^0.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
